Replace deprecated color temp attributes
The color attributes `color_temp` and `kelvin` are deprecated and willl be removed in homeassistant `2026.1`.
As they are removed, I'm removing them for the config and replace it by the new attribute `color_temp_kelvin`.

As this is a breaking change (as old attributes won't work anymore), please list it in the release notes.